+++ src/util.ts
@@ -385,18 +385,12 @@ export function getHeightOfRevision(revision: string): number {
 }
 
 import { stringMd5 } from 'pouchdb-md5';
-import { rev as pouchUtilsRev } from 'pouchdb-utils';
 
 /**
  * Creates a revision string that does NOT include the revision height
  * Copied and adapted from pouchdb-utils/src/rev.js
- * TODO not longer needed when this PR is merged: https://github.com/pouchdb/pouchdb/pull/8274
  */
-export function createRevision(docData: any, deterministic_revs: boolean): string {
-    if (!deterministic_revs) {
-        return pouchUtilsRev(docData, false);
-    }
-
+export function createRevision(docData: any): string {
     const docWithoutRev = Object.assign({}, docData, {
         _rev: undefined,
         _rev_tree: undefined